requirements.txt
mss>=9.0.0
pystray>=0.19.5
pynput>=1.7.6
Pillow>=10.0.0
numpy>=1.24.0
rapidocr-onnxruntime>=1.3.0
# Windows: pip install onnxruntime-directml (see pyproject markers). Linux/mac: onnxruntime.
onnxruntime-directml>=1.20.0,<1.21.0
deep-translator>=1.11.4
